+++ mod/src/RPSoulJar.lua
@@ -75,24 +75,9 @@ function RPSoulJar:SpriteDisplay()
     RPSoulJar.sprites.barLines:Play("BarOverlay12", true) -- This is custom replaced with an 8 charge bar
   end
 
-  local heartLength = 12
-  if maxHearts > 12 then
-    maxHearts = 12 -- After 6 heart containers, it wraps to a second row
-  end
-  local barOffset = (maxHearts / 2) * heartLength
-  if extraLives > 9 then
-    barOffset = barOffset + 20
-    if player:HasCollectible(CollectibleType.COLLECTIBLE_GUPPYS_COLLAR) then -- 212
-      barOffset = barOffset + 6
-    end
-  elseif extraLives > 0 then
-    barOffset = barOffset + 16
-    if player:HasCollectible(CollectibleType.COLLECTIBLE_GUPPYS_COLLAR) then -- 212
-      barOffset = barOffset + 4
-    end
-  end
-
-  local barX = 49 + barOffset -- Place to the right of the red heart containers
+  -- Place the bar to the right of the heart containers
+  -- (which will depend on how many heart containers we have)
+  local barX = 49 + RPSoulJar:GetHeartXOffset()
   local barY = 17
   local barVector = Vector(barX, barY)
 
@@ -109,6 +94,37 @@ function RPSoulJar:SpriteDisplay()
   -- Draw the charge bar 3/3 (the segment lines on top)
   RPSoulJar.sprites.barLines:Render(barVector, Vector(0, 0), Vector(0, 0))
   RPSoulJar.sprites.barLines:Update()
+end
+
+function RPSoulJar:GetHeartXOffset()
+  -- Local variables
+  local game = Game()
+  local player = game:GetPlayer(0)
+  local maxHearts = player:GetMaxHearts()
+  local soulHearts = player:GetSoulHearts()
+  local extraLives = player:GetExtraLives()
+
+  local heartLength = 12 -- This is how long each heart is on the UI in the upper left hand corner
+  -- (this is not in pixels, but in draw coordinates; you can see that it is 13 pixels wide in the "ui_hearts.png" file)
+  local combinedHearts = maxHearts + soulHearts
+  if combinedHearts > 12 then
+    combinedHearts = 12 -- After 6 hearts, it wraps to a second row
+  end
+
+  local offset = (combinedHearts / 2) * heartLength
+  if extraLives > 9 then
+    offset = offset + 20
+    if player:HasCollectible(CollectibleType.COLLECTIBLE_GUPPYS_COLLAR) then -- 212
+      offset = offset + 6
+    end
+  elseif extraLives > 0 then
+    offset = offset + 16
+    if player:HasCollectible(CollectibleType.COLLECTIBLE_GUPPYS_COLLAR) then -- 212
+      offset = offset + 4
+    end
+  end
+
+  return offset
 end
 
 return RPSoulJar